delete duplicate code
